[US_TN] Remove c4 category for compliant reporting and sentencing bug fixes (Recidiviz/recidiviz-data#18548)
## Description of the change

This PR does 2 main things:
1) Based on a request from TN, we're removing the "C4" category which
automatically classifies people on ISC as eligible for compliant
reporting if they were assessed to be on minimum supervision level upon
intake. By removing the `CASE WHEN` that classifies c4 people, they will
now be assessed on all the same requirements as others. We see a slight
increase in C1/C2 (~10-15 people) as a result of this but the vast
majority of formerly c4 folks are not considered eligible anymore

2) Fixes a few more bugs in the sentencing logic:
- some clean up of `OffenseDescription` to improve join between TN's
statutes and sentencing data (see Recidiviz/recidiviz-data#18632 for the upstream fix for this)
- bug in how we were handling ineligible offenses that expired 10+ years
ago
- bug in how we were handling null `Prior` offenses

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f8851be4729112e843a8ca215b04732b025ae21f