[TES] Drop duplicate task names (Recidiviz/recidiviz-data#18700)
## Description of the change

Small fix to only pull unique `task_name` and `completion_event_type`
pairs so that the mapping table goes from 18 rows to 7. Duplicates were
introduced to the original view for opportunities in US_ID and US_IX
that have the same name, along with the 3 states that support
supervision level downgrade.

These duplicates make it so that there are multiple rows in downstream
tables that use this mapping, such as `person_events`.

## Related issues

Closes Recidiviz/recidiviz-data#17978

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 27fe6443c5a91521a3fef205b789c5f80b167c7d