[US_IX] handle lowercase sentence metadata keys (Recidiviz/recidiviz-data#18732)
## Description of the change

Update the `us_ix_consecutive_sentences_preprocessed` view to handle
lowercase or uppercase keys within the `sentence_metadata` JSON blob.

I have tested this logic on the staging and prod data. With this update
the staging version of the view is no longer empty, and the prod view
output is the same.

## Related issues

Closes Recidiviz/recidiviz-data#18726

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 45b32d49bb931d5b2e75b3630067be66231ff9f8