Deprecate compartment sessions unnested (Recidiviz/recidiviz-data#18788)
## Description of the change

This PR deprecates `compartment_sessions_unnested` from recidiviz-data,
following recidiviz/lookerRecidiviz/recidiviz-data#432 which removes all dependencies on this
view in Looker. This includes deleting a terraform deletion so that this
view is no longer added to `unmanaged_views`

After running `git grep` on "compartment_sessions_unnested" there is
still a reference on
[L8936](https://github.com/Recidiviz/recidiviz-data/blob/048a6ee58796c139e9df6af0c6128c5389ef8b74/bandit/bandit-baseline.json#L8936)
of this json (`bandit-baseline.json`). @ageiduschek should that that
reference need to be deleted as well?

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 76967933f00d5b1baa011927cfa62d684e277a7e