[US_MO] Adding in MO query edits for 2-19 transfer (Recidiviz/recidiviz-data#18722)
## Description of the change

A few things are going on that we need to make edits for in the MO query
as well as pull some data ad hoc.

1. Supervision Level and Supervision Level Raw Text are currently null
in MO because we are missing a comma in the raw data table that we get
that information from. This PR adds in that comma after the DOC column.

2. This PR adds in the incremental pulls for TAK047, TAK204, and TAK071.

3. This PR makes TAK017 a full historical due to data deletion causing
issues for Workflows opportunity.

4. In addition to these edits, we need to do the following ad hoc pulls.

- [x] Pull a full historical of LBAKRDTA_TAK047
- [x] Pull a full historical of LBAKRDTA_TAK204
- [x] Pull _either_ a full historical OR an incremental pull from
2022-11-16 of LBAKRDTA_TAK071
- [x] Pull a full historical of LBAKRDTA_TAK039 to correct null ID
errors because of missing comma.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#18635

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8e39de192c33aa8e6ed656caf6582e9d2d89b2a1