[Ingest] Remove include_undocumented_columns infra for raw data views (Recidiviz/recidiviz-data#18876)
## Description of the change

We don't use `include_undocumented_columns=True` except in the context
of a unittest - removing this complexity from our raw data view
generation logic.

There is really no need for this anymore - if an IE/DA wants to use an
undocumented column, they can document it and deploy the newest version
of the latest view to a sandbox.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#18282

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e4daef39a6b692114a5642b35fb75396c01a2882