Setting up safeguards and handling edge cases in SFTP DAG (Recidiviz/recidiviz-data#18877)
## Description of the change

This PR handles a lot of edge cases with the SFTP DAG. The series of
changes include:

- Added `gather_discovered_{remote|ingest_ready}_files` in order to
handle prior discovered, but not yet downloaded/uploaded files from
prior runs. This also includes the files that we just marked discovered
too.
- The download / upload steps were updated to handle the limited
metadata that the `gather` steps send. For instance, the upload step can
parse the timestamp just from the `post_processed_normalized_file_name`
itself.
- The post process downloads step is now moved to before the mark remote
files downloaded step because we want to only mark everything successful
if it's been post processed correctly too. Also, we have a trigger rule
for ALL_DONE because we want to make all of the downloads happen,
regardless of status in order to ensure that we can handle all of the
downloaded files that were successful. If the downloads all fail, the
post processed step will be skipped.
- Added a
`check_{remote|ingest_ready}_files_{downloaded_and_post_processed|uploaded}`
step before the `mark_...downloaded|uploaded` steps which check if the
XCOMS output is None or empty. If so, we skip the mark step.
- As a result of all of the trigger rules that were set in various
steps, I've added comments for what to expect from xcom_pull and have
simplified some of the implementations as a result. I still need to
figure out how to better unit test this (Airflow unfortunately does not
have good E2E integration test frameworks).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#18414

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 75c6a708de06ba60846baa73830fa5d151eabf34