Switching all SFTP over to the environment buckets and DBs (Recidiviz/recidiviz-data#18940)
## Description of the change

This updates the buckets and operations DB connection IDs to use the
real connection IDs (which have been loaded in secret manager) to the
actual ones used in the environment. It also sets up the cron to be
hourly. The current config is empty, so no double-reading/writing
operations will occur unless we manually switch states over. See [launch
plan](https://paper.dropbox.com/doc/Design-Doc-SFTP-in-Airflow--BzLK5byurAAEEXZK6~s7_dYRAg-3tQbqtgmGXO7w9sBVnzJH).

## Type of change

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f93fe0fc0dbf40acc47c0fd6a33c5360a6b199a2