[US_TN] Ingest is_sex_offense for Diversion and ISCSentence tables (Recidiviz/recidiviz-data#18914)
## Description of the change

Currently, the SentencesChargesAndCourtCases_v2 view is only able to
hydrate `is_sex_offense` for data coming from the Sentence table,
resulting in null `is_sex_offense` flags for data coming from the
Diversion/ISCSentence tables. This change updates the view to use
`OffenderStatute` to identify sex offenses in the Diversion table (the
current behavior for the Sentence table), and identifies sex offenses
via pattern matching in the ISCSentence table.

Verified that is_sex_offense will be ingested for the Diversion and ISC
tables by counting SexOffenderFlag over each data source using the
updated view query:
<img width="623" alt="Screen Shot 2023-02-22 at 10 08 15 AM"
src="https://user-images.githubusercontent.com/113454156/220686460-2f480ef2-3702-45f1-8ebe-e51751f69d4f.png">

The only remaining Null values (of which there are just 2) are for the
Sentence table, which was unaffected by this change.

A partial rerun will be needed for this change to take effect on
existing data, which will be initiated once a fix for Recidiviz/recidiviz-data#18632 is deployed
as well.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#18509

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3bdf6b3840b83bf872492fa36e133fb76edb03ca